+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable, and occasionally less notable changes to this project will be docum
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.6]
+- Searchable trait now a blanket implementation for Display and uses Cow
+- removed PrimitiveGroup in favour of is_numeric, is_displayable
+
 ## [0.4.5]
 - grouped data, store, response and executor under data/
 - removed generics from queries, and put them onto the executor instead